% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admit}
\alias{admit}
\title{Graduate admissions}
\format{A data frame with 500 rows and 5 variables:
\describe{
  \item{X}{Case number}
  \item{admit}{Admission decision. 1 = Yes, 2 = No}
  \item{gre}{GRE score}
  \item{gpa}{College GPA}
  \item{rank}{Ranking based on letters of rec. 1-4}
  }
  @source \url{https://osf.io/9tgxm/}}
\usage{
admit
}
\description{
A dataset addressing admission to graduate school
}
\keyword{datasets}
