% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logistic2}
\alias{logistic2}
\title{Compliance with mammography}
\format{A data frame with 164 rows and 8 variables:
\describe{
  \item{case}{Case number}
  \item{comply}{Completed screening? coded as 0, 1}
  \item{physrec}{Had a physicians recommendation? coded as 0, 1}
  \item{knowledg}{Knowledge of mammography}
  \item{benefits}{Perceived benefits of mammography}
  \item{barriers}{Perceived barriers to mammography}
  \item{comply_F}{Factor version of Completed screening? Coded as Yes, No}
  \item{physrec_F}{Factor version of physicians recommendation? Coded as Yes, No}
  }
  @source \url{https://osf.io/9tgxm/}}
\usage{
logistic2
}
\description{
A dataset addressing compliance with mammography screening.
}
\keyword{datasets}
