% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med}
\alias{med}
\title{Mediation Example II}
\format{A data frame with 543 rows and 8 variables:
\describe{
  \item{id}{ID}
  \item{negatt}{negatt - not used in example}
  \item{stereo}{Stereotype endorsement}
  \item{anxiety}{Intergroup anxiety}
  \item{symbolic}{Symbolic Threat}
  \item{realistic}{Realistic Threat}
  \item{quality}{Contact Quality}
  \item{att}{Attitudes toward outgroup}
  }}
\usage{
med
}
\description{
A dataset with measures of contact, threats, and prejudice (with some missing data)
}
\keyword{datasets}
