% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mednomiss}
\alias{mednomiss}
\title{Mediation Example I}
\format{A data frame with 534 rows and 8 variables:
\describe{
  \item{id}{ID}
  \item{negatt}{negatt - not used in example}
  \item{stereo}{Stereotype endorsement}
  \item{anxiety}{Intergroup anxiety}
  \item{symbolic}{Symbolic Threat}
  \item{realistic}{Realistic Threat}
  \item{quality}{Contact Quality}
  \item{att}{Attitudes toward outgroup}
  }}
\usage{
mednomiss
}
\description{
A dataset with measures of contact, threats, and prejudice
}
\keyword{datasets}
