% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mindful}
\alias{Mindful}
\title{Mindfulness data}
\format{A data frame with 331 rows and 7 variables:
\describe{
  \item{SelfAware}{Self Awareness}
  \item{Mindfulness}{Mindfulness}
  \item{LifeSat}{Life Satisfaction}
  \item{OBE}{Observing}
  \item{DES}{Describing}
  \item{JUD}{Acting with Awareness}
  \item{ACT}{Non Judgemental}
  }}
\usage{
Mindful
}
\description{
A dataset with measures of mindfulness, life satisfaction, and self awareness
}
\keyword{datasets}
