\name{mvstats-package}
\alias{mvstats-package}
\alias{mvstats}
\docType{package}
\title{
\packageTitle{mvstats}
}
\description{
\packageDescription{mvstats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mvstats}
\packageIndices{mvstats}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mvstats}

Maintainer: \packageMaintainer{mvstats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
